fix: move to adjusting transforms' z-component
The old way of modifying the `ExtractedSprite`s doesn't work starting
from 0.12(!) since the `ExtractedSprite` has no way of going back to the
original entity. I'd like to avoid tying this plugin too closely to
Bevy internals, so the simplest thing to do is to just ensure that any
entity with a sprite layer will wind up with the z-coordinate set
properly.

This also means that entities that are children of something with a
sprite layer will inherit its layer, which is a good thing.